# swaggerui/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Simple Inventory API",
    "description": "This is a simple API",
    "contact": {
      "email": "you@your-company.com"
    },
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    },
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "https://virtserver.swaggerhub.com/ALIASADDIK_1/TODO/1.0.0",
      "description": "SwaggerHub API Auto Mocking"
    }
  ],
  "paths": {
    "/": {
      "get": {
        "summary": "shows the list",
        "description": "the main page shows the user all their tasks",
        "operationId": "showList",
        "responses": {
          "200": {
            "description": "search results matching criteria",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Task"
                  }
                }
              }
            }
          }
        }
      },
      "put": {
        "summary": "Edits a task",
        "description": "edits a task already in the system",
        "operationId": "editTask",
        "parameters": [
          {
            "name": "editTitle",
            "in": "query",
            "required": false,
            "style": "form",
            "explode": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "editDone",
            "in": "query",
            "required": false,
            "style": "form",
            "explode": true,
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "id",
            "in": "query",
            "required": false,
            "style": "form",
            "explode": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "description": "new Item to add",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Task"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "item updated successfully"
          },
          "400": {
            "description": "invalid input, object invalid"
          }
        }
      },
      "post": {
        "summary": "adds a task",
        "description": "Adds a task to the system",
        "operationId": "addTask",
        "requestBody": {
          "description": "new Item to add",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Task"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "item created"
          },
          "400": {
            "description": "invalid input, object invalid"
          },
          "409": {
            "description": "an existing item already exists"
          }
        }
      },
      "delete": {
        "summary": "Deletes a task",
        "description": "edits a task already in the system",
        "operationId": "deleteTask",
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "required": true,
            "style": "form",
            "explode": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "item deleted successfully"
          },
          "400": {
            "description": "invalid input, object invalid"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Task": {
        "required": [
          "done",
          "id",
          "title"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "done": {
            "type": "boolean"
          },
          "title": {
            "type": "string"
          }
        }
      }
    }
  }
}